requirements.txt
acres==0.5.0
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
easyocr==1.7.2
etelemetry==0.3.1
executing==2.2.0
filelock==3.19.1
fitz==0.0.1.dev2
fonttools==4.59.2
fsspec==2025.9.0
hf-xet==1.1.9
httplib2==0.30.0
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy_loader==0.4
looseversion==1.3.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
prov==2.1.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyclipper==1.3.0.post6
pycocotools==2.0.10
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.1
rdflib==7.1.4
regex==2025.9.1
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
setuptools==75.8.0
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tifffile==2025.8.28
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.56.0
triton==3.4.0
typer==0.16.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
